requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.16; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
fastapi==0.108.0; python_version >= '3.8'
frozenlist==1.6.0; python_version >= '3.9'
google-ai-generativelanguage==0.6.4; python_version >= '3.7'
google-api-core[grpc]==2.24.2; python_version >= '3.7'
google-api-python-client==2.167.0; python_version >= '3.7'
google-auth==2.39.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1; python_version >= '3.6'
google-generativeai==0.5.4; python_version >= '3.9'
googleapis-common-protos==1.70.0; python_version >= '3.7'
grpcio==1.71.0; python_version >= '3.9'
grpcio-status==1.62.3; python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.8; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx==0.27.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.1.10; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-community==0.0.38; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-core==0.1.53; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-google-genai==1.0.3; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.0.2; python_full_version >= '3.8.1' and python_version < '4.0'
langserve==0.1.1; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
langsmith==0.1.147; python_full_version >= '3.8.1' and python_version < '4.0'
marshmallow==3.26.1; python_version >= '3.9'
multidict==6.4.3; python_version >= '3.9'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
orjson==3.10.16; python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
propcache==0.3.1; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==4.25.6; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pydantic==2.11.3; python_version >= '3.9'
pydantic-core==2.33.1; python_version >= '3.9'
pyparsing==3.2.3; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.40; python_version >= '3.7'
sse-starlette==2.1.0; python_version >= '3.8'
starlette==0.32.0.post1; python_version >= '3.8'
tenacity==8.5.0; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspect==0.9.0
typing-inspection==0.4.0; python_version >= '3.9'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.4.0; python_version >= '3.9'
uvicorn==0.23.2; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
yarl==1.20.0; python_version >= '3.9'
